pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "filldoc"
version = "0.1.0"
description = "Desktop app for filling Word templates from Excel project cards"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "PyMuPDF>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
]

[project.scripts]
filldoc = "filldoc.app:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
